bookmaker/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
ansicon==1.89.0
appdirs==1.4.4
arrow==1.4.0
asgiref==3.11.0
async-timeout==5.0.1
asyncio==4.0.0
attrs==25.4.0
base58==2.1.1
beautifulsoup4==4.14.2
billiard==4.2.4
bip32utils==0.3.post4
bip_utils==2.10.0
blessed==1.27.0
blinker==1.9.0
brotli==1.2.0
bs4==0.0.2
cbor2==5.7.1
celery-types==0.24.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coincurve==21.0.0
colorama==0.4.6
crcmod==1.7
cron_descriptor==2.0.6
cryptography==46.0.4
cssselect==1.3.0
decorator==5.2.1
Django==5.2.8
django-celery-beat==2.8.1
django-picklefield==3.4.0
django-q2==1.9.0
django-timedeltafield==0.7.10
django-timezone==0.3
django-timezone-field==7.2.1
ecdsa==0.19.1
ed25519-blake2b==1.4.1
fake-useragent==2.2.0
frozenlist==1.8.0
greenlet==3.0.3
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
jinxed==1.3.0
kaitaistruct==0.11
kombu==5.6.2
lxml==6.0.2
matches==0.0.1
mnemonic==0.21
multidict==6.7.0
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
pillow==12.0.0
playwright==1.44.0
playwright-stealth==2.0.2
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
py-sr25519-bindings==0.2.3
pyasn1==0.6.2
pycountry==24.6.1
pycparser==2.23
pycryptodome==3.23.0
pydivert==2.1.0
pyee==11.1.0
PyNaCl==1.6.1
pyOpenSSL==25.3.0
pyparsing==3.3.2
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pproxy==2.7.9
qrcode==8.2
requests==2.32.5
requests-html==0.10.0
selenium==4.9.1
selenium-wire==5.1.0
self==2020.12.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.3
tqdm==4.67.1
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
undetected-chromedriver==3.5.5
urllib3==1.26.20
vine==5.1.0
w3lib==2.3.1
wcwidth==0.2.14
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==10.4
whitenoise==6.11.0
wsproto==1.3.2
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0